netbox_create_data/core/requirements.txt
awesome-slugify==1.6.5
et-xmlfile==1.1.0
excel2json-3==0.1.6
Flask==2.0.1
Flask-SocketIO==5.0.3
Jinja2==3.0.1
openpyxl==3.0.7
pandas==1.1.5
pynetbox==6.1.2
python-engineio==4.2.0
python-socketio==5.3.0
PyYAML==3.12
regex==2021.4.4
requests==2.25.1
Unidecode==0.4.21
Werkzeug==2.0.1
xlrd==2.0.1